backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # If using PostgreSQL later

# Task Queue
celery==5.3.4
redis==5.0.1

# AI/ML
anthropic==0.7.7
sentence-transformers==2.2.2
torch==2.1.1  # For semantic matching
numpy==1.24.3

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Web Scraping
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
httpx==0.25.2
webdriver-manager==4.0.1

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
reportlab==4.0.7
jinja2==3.1.2

# Email
sendgrid==6.11.0  # Alternative to Gmail API

# Data Processing
pandas==2.1.3
openpyxl==3.1.2  # For Excel
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.38.0  # Optional: error tracking

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
